+++ src/components/modal/index.js
@@ -1,16 +1,18 @@
 import Vue from 'vue'
 import Modal from './Modal'
 
+
+export { Modal }
 export default {
     open(params) {
-        let propsData
-        if (typeof params === 'string') {
-            propsData = {
-                content: params
-            }
-        } else {
-            propsData = params
+        let content
+        if (typeof params === 'string') content = params
+
+        const defaultParam = {
+            programmatic: true,
+            content
         }
+        const propsData = Object.assign(defaultParam, params)
 
         const ModalComponent = Vue.extend(Modal)
         return new ModalComponent({