+++ src/index.js
@@ -13,7 +13,7 @@ import Switch from './components/switch'
 import Tooltip from './components/tooltip'
 
 import Dialog from './components/dialog'
-import Modal from './components/modal'
+import ModalProgrammatic, { Modal } from './components/modal'
 import Snackbar from './components/snackbar'
 import Toast from './components/toast'
 
@@ -28,6 +28,7 @@ const components = {
     Icon,
     Input,
     Message,
+    Modal,
     Notification,
     Pagination,
     Radio,
@@ -53,7 +54,7 @@ components.install = (Vue, options = {}) => {
     }
 
     Vue.prototype.$snackbar = Snackbar
-    Vue.prototype.$modal = Modal
+    Vue.prototype.$modal = ModalProgrammatic
     Vue.prototype.$toast = Toast
     Vue.prototype.$dialog = Dialog
 }